+++ src/Ntfs/LZNT1.cs
@@ -166,7 +166,7 @@ namespace DiscUtils.Ntfs
                     // Set the header to indicate non-compressed block
                     Utilities.WriteBytesLittleEndian((ushort)(0x3000 | (BlockSize - 1)), compressed, compressedOffset + (int)headerPosition);
 
-                    Array.Copy(source, (int)sourceOffset + sourceCurrentBlock, compressed, compressedOffset + headerPosition + 2, BlockSize);
+                    Array.Copy(source, (int)(sourceOffset + sourceCurrentBlock), compressed, (int)(compressedOffset + headerPosition + 2), BlockSize);
                     destPointer = (uint)(headerPosition + 2 + BlockSize);
 
                     // Make sure decompression stops by setting the next two bytes to null, prevents us from having to
